src/core/scanner.py
# src/core/scanner.py
from scapy.all import *
import nmap
from typing import Dict, List, Optional
import asyncio

class PortScanner:
    def __init__(self):
        self.nm = nmap.PortScanner()
    
    async def scan_ports(self, target_ip: str, ports: str = "1-1000") -> Dict:
        """
        Perform a port scan on the target IP
        """
        try:
            # Run nmap scan
            self.nm.scan(target_ip, ports, arguments="-sV -O")
            
            results = {
                "target": target_ip,
                "ports": [],
                "os_match": self.nm[target_ip].get('osmatch', []),
                "status": "completed"
            }
            
            # Process port results
            for proto in self.nm[target_ip].all_protocols():
                ports = self.nm[target_ip][proto].keys()
                for port in ports:
                    port_info = self.nm[target_ip][proto][port]
                    results["ports"].append({
                        "port": port,
                        "state": port_info["state"],
                        "service": port_info["name"],
                        "version": port_info.get("version", "")
                    })
            
            return results
            
        except Exception as e:
            return {
                "target": target_ip,
                "status": "failed",
                "error": str(e)
            }

class WebScanner:
    async def scan_web_vulnerabilities(self, target_url: str) -> Dict:
        """
        Scan web application for common vulnerabilities
        """
        # This is a placeholder - you'll need to implement actual web scanning logic
        vulnerabilities = []
        
        # Test for XSS
        xss_vulns = await self._test_xss(target_url)
        vulnerabilities.extend(xss_vulns)
        
        # Test for SQL Injection
        sql_vulns = await self._test_sqli(target_url)
        vulnerabilities.extend(sql_vulns)
        
        return {
            "target_url": target_url,
            "vulnerabilities": vulnerabilities,
            "status": "completed"
        }

    async def _test_xss(self, url: str) -> List[Dict]:
        # Implement XSS testing logic
        pass

    async def _test_sqli(self, url: str) -> List[Dict]:
        # Implement SQL Injection testing logic
        pass